update:Dish 模块中，批量删除菜品 (delete 方法) 完善

+++ sky-take-out/sky-server/src/main/java/com/sky/service/DishService.java
@@ -21,5 +21,5 @@ public interface DishService {
 
     void startOrStop(Integer status, Integer id);
 
-    void delete(List<Integer> ids);
+    void delete(List<Long> ids);
 }
